Added Subfolder responses_bot
