+++ responses_bot/settings.json
@@ -1,0 +1,6 @@
+{
+    "token" : "TOKEN_HERE",
+    "owner" : 123456,
+    "prefix" : "PREFIX_HERE",
+    "ignore_bot" : true
+}
