--- settings.json
@@ -1,6 +1,0 @@
-{
-    "token" : "TOKEN_HERE",
-    "owner" : 123456,
-    "prefix" : "PREFIX_HERE",
-    "ignore_bot" : true
-}
